% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADTGP.R
\name{ADTGP}
\alias{ADTGP}
\title{Fits model and draws posterior samples of ADT counts}
\usage{
ADTGP(
  igg = NULL,
  poi = NULL,
  design_matrix = NULL,
  priors = list(mu0 = "gamma(7, 2)", phi = "gamma(.5, .5)", etasq = "normal(2, 1)", rhosq
    = "uniform(0, 5000)"),
  fn = "StanModel_fit_data.stan",
  parallel_chains = 4,
  iter_warmup = 5000,
  iter_sampling = 5000
)
}
\arguments{
\item{igg}{A numeric vector giving the igg count.}

\item{poi}{A numeric vector giving the protein of interest count.}

\item{design_matrix}{A design matrix.}

\item{priors}{Likelihood functions in stan syntax as a named
list of characters. See reference for avaiable options.
\describe{
  \item{\strong{mu0: the grand mean in log scale}}
  {Default is "gamma(7, 2)"}
  \item{\strong{phi: overdispersion}}
  {Default is "gamma(.5, .5)"}
  \item{\strong{etasq: maximum covariance between any two cells}}
  {Default is "normal(2, 1)"}
  \item{\strong{rhosq: rate of decline in covariance}}
  {Default is "uniform(0, 5000)"}
}}

\item{fn}{File path to store stan model file.}

\item{parallel_chains}{Number of parallel chains to run. Defaults to 4.}

\item{iter_warmup}{Number of warmup iterations before sampling.
Defaults to 5,000.}

\item{iter_sampling}{Number of samples to draw from each chain.
Defaults to 5,000.}

\item{variables}{A character vector declaring which sampled variables
to return.
Defaults to "poi_rep" which is the count of the
protein of interest.}
}
\value{
A stanfit object and a matrix of draws stored in a named list.
}
\description{
Writes a stan model and draws samples from the posterior distribution of
protein count given a vector of isotype control (igg) count.
}
\references{
\url{https://mc-stan.org/docs/reference-manual/}
}
