% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_example.R
\name{ADTGP_RunSimulation}
\alias{ADTGP_RunSimulation}
\title{Runs the simulated example in the manuscript}
\usage{
ADTGP_RunSimulation()
}
\description{
Returns the true dataframe, the noisy dataframe, and the stanfit object.
Also draws two plots in the working directory as a side effect.
}
