% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{idh_pdx}
\alias{idh_pdx}
\title{Raw protein data from isozyme switching dataset}
\format{
A dataframe with 47 columns and 200 rows.
}
\usage{
idh_pdx
}
\description{
A dataset containing raw counts of 46 proteins from 200 cells.
An IDH1-mutated AML sample was mixed with an IDH2-mutated AML sample
at a 1-to-1 ratio. The mixed cells were injected into the tail vein
 of immunodeficient mice to generate mixed patient derived xenografts.
 The animals were treated with vehicle control or ivosidenib,
 an IDH1 inhibitor, to induce the differentiation of AML cells for eight
 weeks. At endpoint, bone marrow cells were harvested from the
 animals (n=5 per treatment) and processed through MissionBio's
 Tapestri single-cell multi-omics pipeline.
}
\details{
veh and ivo are short-hand for vehicle and ivosidenib, respectively.
}
\keyword{datasets}
