% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tenx_pbmc10k}
\alias{tenx_pbmc10k}
\title{Raw protein data from 10X Genomic's PBMC 10K dataset}
\format{
A matrix with 17 columns and 7,865 rows.
#' @source \url{https://www.10xgenomics.com/datasets/10-k-pbm-cs-from-a-healthy-donor-gene-expression-and-cell-surface-protein-3-standard-3-0-0}
}
\usage{
tenx_pbmc10k
}
\description{
A dataset containing raw counts of 17 proteins from 7,865 cells of a healthy donor.
}
\keyword{datasets}
